requirements/base.txt
anyio==4.7.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
click==8.1.7
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.6.0
django-stubs==4.2.7
django-stubs-ext==5.1.1
exceptiongroup==1.2.2
graphql-core==3.2.5
gunicorn==23.0.0
h11==0.14.0
idna==3.10
libcst==1.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.7.1
mypy-extensions==1.0.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.42.0
tomli==2.2.1
typer==0.15.1
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
whitenoise==6.8.2
django-two-factor-auth
django-two-factor-auth[phonenumbers]
django-mathfilters
fillpdf==0.7.3
requests